Completed Chapter 10 challenge.
i set a member variable `mChangedItemPosition` on the adapter, created a getter and setter for said variable. I set `mChangedItemPosition` to `getAdapterPosition` in the `onClick` method that is set on the view holder in the ViewHolder class. then, i call the fragment's `updateUI()` method, in which i check for the adapter, if no adapter exists, I create a new one. If an adapter exists, I get the mChangedItemPosition and then pass that to the notifyItemChanged(position)